SelectableModelsController mixin
